+++ tests_integration/spcs/test_image_repository.py
@@ -1,6 +1,8 @@
 import pytest
+from snowflake.cli.api.project.util import escape_like_pattern
 
-from tests_integration.test_utils import contains_row_with
+from tests_integration.test_utils import contains_row_with, row_from_snowflake_session
+from tests_integration.testing_utils.naming_utils import ObjectNameProvider
 
 INTEGRATION_DATABASE = "SNOWCLI_DB"
 INTEGRATION_SCHEMA = "PUBLIC"
@@ -9,6 +11,7 @@ INTEGRATION_REPOSITORY = "snowcli_repository"
 
 @pytest.mark.integration
 def test_list_images_tags(runner):
+    # test assumes the testing environment has been set up with /SNOWCLI_DB/PUBLIC/snowcli_repository/snowpark_test:1
     _list_images(runner)
     _list_tags(runner)
 
@@ -36,7 +39,6 @@ def _list_images(runner):
     )
 
 
-@pytest.mark.integration
 def _list_tags(runner):
     result = runner.invoke_with_connection_json(
         [
@@ -59,3 +61,23 @@ def _list_tags(runner):
             "tag": f"/{INTEGRATION_DATABASE}/{INTEGRATION_SCHEMA}/{INTEGRATION_REPOSITORY}/snowpark_test:1"
         },
     )
+
+
+@pytest.mark.integration
+def test_get_repo_url(runner, snowflake_session, test_database):
+    repo_name = ObjectNameProvider("TEST_REPO").create_and_get_next_object_name()
+    snowflake_session.execute_string(f"create image repository {repo_name}")
+
+    created_repo = snowflake_session.execute_string(
+        f"show image repositories like '{escape_like_pattern(repo_name)}'"
+    )
+    created_row = row_from_snowflake_session(created_repo)[0]
+    created_name = created_row["name"]
+    assert created_name == repo_name.upper()
+
+    expect_url = created_row["repository_url"]
+    result = runner.invoke_with_connection(
+        ["spcs", "image-repository", "url", created_name]
+    )
+    assert isinstance(result.output, str), result.output
+    assert result.output.strip() == expect_url